cuda gemm

+++ src/cuda/dgemm/safe_memory.h
@@ -7,9 +7,8 @@
 
 #ifndef SAFE_MEMORY_H_
 #define SAFE_MEMORY_H_
-#include <cstdlib>
 
-void* safe_malloc(size_t size);
-int safe_cuda_malloc_cover(void **ptr, size_t size);
+void* safe_malloc(unsigned long size);
+int safe_cuda_malloc_cover(void **ptr, unsigned long size);
 
 #endif /* SAFE_MEMORY_H_ */
